requirements.txt
asgiref==3.9.0
Django==5.0.1
gunicorn==21.2.0
packaging==25.0
psycopg2-binary
setuptools==78.1.1
sqlparse==0.5.3
wheel==0.45.1
openai
django-environ
requests
whitenoise